requirements.txt
# CertifyOS Jira Automation Requirements
# Install with: pip install -r requirements.txt

# Core Jira API client
atlassian-python-api==3.41.5

# HTTP requests
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0

# Date/time utilities
python-dateutil>=2.8.2

# Optional: For Google Sheets integration
gspread>=5.12.4
google-auth>=2.25.2
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# Optional: For enhanced CLI experience
rich>=13.0.0
click>=8.1.0

# Optional: For configuration management
pyyaml>=6.0.1

# Optional: For data validation
pydantic>=2.0.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-mock>=3.11.0
black>=23.0.0
flake8>=6.0.0
